refactor: keyword 기반 식당 조회 기능을 RestaurantReviewStatService로 이동
해당 조회 기능은 평균 별점과 리뷰 개수 같은 집계 데이터를 동반한 조회 기능이기 때문에 RestaurantReviewStatService를 통해 RestaurantReviewStatRepository로 READ-DB로부터 데이터를 읽어와야 함.
