+++ src/main/java/com/jvnlee/catchdining/domain/restaurant/controller/RestaurantController.java
@@ -6,6 +6,7 @@ import com.jvnlee.catchdining.domain.restaurant.dto.RestaurantSearchResponseDto;
 import com.jvnlee.catchdining.domain.restaurant.dto.RestaurantSearchRequestDto;
 import com.jvnlee.catchdining.domain.restaurant.dto.RestaurantViewDto;
 import com.jvnlee.catchdining.domain.restaurant.model.SortBy;
+import com.jvnlee.catchdining.domain.restaurant.service.RestaurantReviewStatService;
 import com.jvnlee.catchdining.domain.restaurant.service.RestaurantService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -20,6 +21,8 @@ public class RestaurantController {
 
     private final RestaurantService restaurantService;
 
+    private final RestaurantReviewStatService restaurantReviewStatService;
+
     @PostMapping
     @PreAuthorize("hasRole('ROLE_OWNER')")
     public Response<Void> register(@RequestBody RestaurantDto restaurantDto) {
@@ -31,7 +34,7 @@ public class RestaurantController {
     public Response<Page<RestaurantSearchResponseDto>> search(@RequestParam String keyword,
                                                             @RequestParam(required = false, defaultValue = "none") SortBy sortBy,
                                                             Pageable pageable) {
-        Page<RestaurantSearchResponseDto> data = restaurantService
+        Page<RestaurantSearchResponseDto> data = restaurantReviewStatService
                 .search(new RestaurantSearchRequestDto(keyword, sortBy, pageable));
         return new Response<>("식당 검색 결과", data);
     }
